[MobileHc-4]: Refactor - relocate all files and folders
