+++ lib/features/auth/bloc/registration_state.dart
