+++ lib/features/auth/ui/register_first_step_widget.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../../component/src.dart';
-import '../viewmodel/src.dart';
+import '../../../shared/components/src.dart';
+import '../bloc/src.dart';
 
 class RegisterFirstStep extends StatefulWidget {
   const RegisterFirstStep({super.key});
