+++ lib/features/auth/ui/register_page.dart
@@ -5,12 +5,12 @@ import 'package:flutter_svg/svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:top_snackbar_flutter/custom_snack_bar.dart';
 import 'package:top_snackbar_flutter/top_snack_bar.dart';
-import '../../component/src.dart';
+import '../../../shared/components/src.dart';
 import '../../routes/route_paths.dart';
-import '../binding/registration_binding.dart';
-import '../viewmodel/registration_bloc.dart';
-import '../viewmodel/registration_event.dart';
-import '../viewmodel/registration_state.dart';
+import '../di/registration_binding.dart';
+import '../bloc/registration_bloc.dart';
+import '../bloc/registration_event.dart';
+import '../bloc/registration_state.dart';
 import 'register_first_step_widget.dart';
 import 'register_last_step_widget.dart';
 import 'register_widgets.dart';
