+++ lib/shared/components/ButtonComponents.dart
@@ -1,7 +1,7 @@
 // ignore: file_names
 import 'package:flutter/material.dart';
-import 'package:flutter_health_care/presentation/component/BaseText.dart';
-import 'package:flutter_health_care/presentation/component/ColorComponent.dart';
+import 'package:flutter_health_care/shared/components/BaseText.dart';
+import 'package:flutter_health_care/shared/components/ColorComponent.dart';
 import 'package:flutter_svg/svg.dart';
 
 class PrimaryButton extends StatelessWidget {
