+++ lib/shared/components/ColorComponent.dart
